+++ crates/vault-agent/src/server.rs
@@ -288,6 +288,7 @@ async fn dispatch(req: Request, state: &Arc<Mutex<AgentState>>) -> Response {
             totp,
             uri,
             card,
+            identity,
         } => {
             let w = crate::state::CipherWrite {
                 name: Some(name),
@@ -298,6 +299,7 @@ async fn dispatch(req: Request, state: &Arc<Mutex<AgentState>>) -> Response {
                 totp,
                 uri,
                 card,
+                identity,
             };
             let mut s = state.lock().await;
             let res = s.add_cipher(cipher_type, w).await;
@@ -318,6 +320,7 @@ async fn dispatch(req: Request, state: &Arc<Mutex<AgentState>>) -> Response {
             totp,
             uri,
             card,
+            identity,
         } => {
             let w = crate::state::CipherWrite {
                 name,
@@ -328,6 +331,7 @@ async fn dispatch(req: Request, state: &Arc<Mutex<AgentState>>) -> Response {
                 totp,
                 uri,
                 card,
+                identity,
             };
             let mut s = state.lock().await;
             let res = s.edit_cipher(&selector, w).await;